+++ sources/Google.Solutions.Apis.Test/Google.Solutions.Apis.Test.csproj
@@ -47,6 +47,9 @@
     <Reference Include="Google.Apis.Auth.PlatformServices, Version=1.57.0.0, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
       <HintPath>..\packages\Google.Apis.Auth.1.57.0\lib\net461\Google.Apis.Auth.PlatformServices.dll</HintPath>
     </Reference>
+    <Reference Include="Google.Apis.CloudResourceManager.v1, Version=1.57.0.2677, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
+      <HintPath>..\packages\Google.Apis.CloudResourceManager.v1.1.57.0.2677\lib\net45\Google.Apis.CloudResourceManager.v1.dll</HintPath>
+    </Reference>
     <Reference Include="Google.Apis.Compute.v1, Version=1.57.0.2714, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
       <HintPath>..\packages\Google.Apis.Compute.v1.1.57.0.2714\lib\net45\Google.Apis.Compute.v1.dll</HintPath>
     </Reference>
@@ -82,7 +85,9 @@
     <Compile Include="Client\TestClientServiceMtlsExtensions.cs" />
     <Compile Include="Client\TestRestClient.cs" />
     <Compile Include="Client\TestUserAgent.cs" />
+    <Compile Include="Compute\TestComputeEngineAdapter.cs" />
     <Compile Include="Compute\TestMetadataExtensions.cs" />
+    <Compile Include="Crm\TestResourceManagerAdapter.cs" />
     <Compile Include="Locator\TestAccessLevelLocator.cs" />
     <Compile Include="Locator\TestDiskTypeLocator.cs" />
     <Compile Include="Locator\TestImageLocator.cs" />
