+++ sources/Google.Solutions.Apis.Test/packages.config
@@ -3,6 +3,7 @@
   <package id="Castle.Core" version="5.0.0" targetFramework="net462" />
   <package id="Google.Apis" version="1.57.0" targetFramework="net462" />
   <package id="Google.Apis.Auth" version="1.57.0" targetFramework="net462" />
+  <package id="Google.Apis.CloudResourceManager.v1" version="1.57.0.2677" targetFramework="net462" />
   <package id="Google.Apis.Compute.v1" version="1.57.0.2714" targetFramework="net462" />
   <package id="Google.Apis.Core" version="1.57.0" targetFramework="net462" />
   <package id="Moq" version="4.18.1" targetFramework="net462" />
