+++ sources/Google.Solutions.Apis/Google.Solutions.Apis.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Compute\IComputeEngineAdapter.cs" />
     <Compile Include="Compute\InstanceExtensions.cs" />
     <Compile Include="Compute\InstanceResourceExtensions.cs" />
+    <Compile Include="Compute\Permissions.cs" />
     <Compile Include="Crm\ResourceManagerAdapter.cs" />
     <Compile Include="Diagnostics\IHelpTopic.cs" />
     <Compile Include="Diagnostics\TraceSourceLogger.cs" />
