+++ sources/Google.Solutions.IapDesktop.Application.Test/Google.Solutions.IapDesktop.Application.Test.csproj
@@ -99,14 +99,12 @@
     <Compile Include="Host\TestInstall.cs" />
     <Compile Include="Host\TestProfile.cs" />
     <Compile Include="ObjectModel\TestServiceAttribute.cs" />
-    <Compile Include="Services\Adapters\TestComputeEngineAdapter.cs" />
     <Compile Include="Services\Adapters\TestGithubAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\Adapters\TestHttpProxyAdapter.cs" />
     <Compile Include="Services\Adapters\TestInstanceExtensions.cs" />
     <Compile Include="Services\TestUpdateService.cs" />
     <Compile Include="Services\Windows\TestWindowsCredentialService.cs" />
-    <Compile Include="Services\Adapters\TestResourceManagerAdapter.cs" />
     <Compile Include="Services\Auth\TestAuthorization.cs" />
     <Compile Include="Services\Auth\TestSignInAdapter.cs" />
     <Compile Include="Services\Integration\TestSessionBroker.cs" />
